+++ phpmyfaq/src/phpMyFAQ/System.php
@@ -62,7 +62,7 @@ class System
     /**
      * Plugin version.
      */
-    private const PLUGIN_VERSION = '0.1.0';
+    private const PLUGIN_VERSION = '0.2.0';
 
     /**
      * Minimum required PHP version.